requirements.txt
requests==2.31.0
beautifulsoup4==4.14.3
Flask==2.2.5
Flask-Cors==5.0.0
lxml==4.9.4
Markdown==3.7
markdownify==1.2.2
numpy==1.24.4
openpyxl==3.1.5
pandas==1.5.3
python-docx==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
tqdm==4.67.1
Werkzeug==3.0.6

blueprint